requirements.txt
--find-links https://download.pytorch.org/whl/cu118/torch_stable.html
numpy==1.23.5
onnx==1.16.0
onnxruntime-gpu==1.16.0; sys_platform == 'win32'
onnxruntime-directml==1.16.0; sys_platform == 'win32'
onnx2pytorch==0.5.0
PySoundFile==0.9.0.post1
torch==2.0.1+cu118; sys_platform != 'darwin'
torch==2.0.1; sys_platform == 'darwin'
osascript==2020.12.3; sys_platform == 'darwin'
PyAudioWPatch==0.2.12.6; sys_platform == 'win32'
pyaudio==0.2.14; sys_platform != 'win32'
librosa==0.9.2
keyboard==0.13.5
comtypes==1.4.8; sys_platform == 'win32'
pycaw==20240210; sys_platform == 'win32'
pyinstaller==6.11.0
